Add separete uploader tool and refactor it to be used throughout file pages. Small refactorings from using fetch to using fetcher
